pyproject.toml
[tool.poetry]
name = "api-psicologia"
version = "0.1.0"
description = "Marketplace de Psicologia — FastAPI + SQLModel + PostgreSQL"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi           = "^0.111.0"
uvicorn           = { extras = ["standard"], version = "^0.30.0" }

# ── Database ───────────────────────────────────────────────────────────────────
sqlmodel          = "^0.0.19"
asyncpg           = "^0.29.0"           # async PostgreSQL driver
alembic           = "^1.13.0"           # database migrations

# ── Configuration ──────────────────────────────────────────────────────────────
pydantic-settings = "^2.3.0"
python-dotenv     = "^1.0.0"

# ── Auth ───────────────────────────────────────────────────────────────────────
python-jose       = { extras = ["cryptography"], version = "^3.3.0" }
passlib           = { extras = ["bcrypt"], version = "^1.7.4" }

# ── Payments ───────────────────────────────────────────────────────────────────
stripe            = "^9.12.0"

# ── Timezone data (required on Windows — zoneinfo has no OS tz on Win) ────────
tzdata            = ">=2024.1"

# ── Video / HTTP ───────────────────────────────────────────────────────────────
httpx             = "^0.27.0"

# ── IA / Transcrição ──────────────────────────────────────────────────────────
anthropic         = "^0.40.0"           # Claude API (resumo clínico)
faster-whisper    = "^1.0.0"            # Transcrição local (LGPD — dado não sai do servidor)
langfuse          = "^2.0.0"            # Observabilidade de LLMs (opcional)
slowapi           = "^0.1.9"            # Rate limiting
apscheduler       = "^3.10.0"          # Scheduler para lembretes de sessão

[tool.poetry.group.dev.dependencies]
# ── Testing ────────────────────────────────────────────────────────────────────
pytest            = "^8.2.0"
pytest-asyncio    = "^0.23.0"
pytest-cov        = "^5.0.0"
httpx             = "^0.27.0"           # AsyncClient for TestClient

# ── Quality ────────────────────────────────────────────────────────────────────
ruff              = "^0.4.0"            # linter + formatter
mypy              = "^1.10.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]   # FastAPI Depends() in default args is intentional

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]

# ── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.12"
strict                 = true
ignore_missing_imports = true
